configure validation

+++ src/main/java/com/mnc/booking/model/MeasurementUnit.java
@@ -3,11 +3,10 @@ package com.mnc.booking.model;
 import lombok.Data;
 
 import javax.persistence.Embeddable;
-import java.math.BigDecimal;
 
 @Data
 @Embeddable
 public class MeasurementUnit {
-  private BigDecimal value;
+  private Integer value;
   private String unit;
 }
